media系の型周りをリファクタ
