+++ frontend/src/api/internal/media/detail.ts
@@ -2,23 +2,22 @@ import { apiClient } from 'lib/axios/internal'
 import { cookieHeader } from 'lib/config'
 import { ApiOut, apiOut } from 'lib/error'
 import { Req } from 'types/global'
-import { Video, Music, Comic, Picture, Chat, BlogDetailOut, FollowIn, FollowOut } from 'types/internal/media'
-import { apiVideo, apiMusic, apiComic, apiPicture, apiBlog, apiChat, apiFollow } from 'api/uri'
-import { camelSnake } from 'utils/functions/convertCase'
+import { VideoDetailOut, MusicDetailOut, ComicDetailOut, PictureDetailOut, BlogDetailOut, ChatDetailOut } from 'types/internal/media/detail'
+import { apiVideo, apiMusic, apiComic, apiPicture, apiBlog, apiChat } from 'api/uri'
 
-export const getVideo = async (id: number, req?: Req): Promise<ApiOut<Video>> => {
+export const getVideo = async (id: number, req?: Req): Promise<ApiOut<VideoDetailOut>> => {
   return await apiOut(apiClient.get(apiVideo(id), cookieHeader(req)))
 }
 
-export const getMusic = async (id: number, req?: Req): Promise<ApiOut<Music>> => {
+export const getMusic = async (id: number, req?: Req): Promise<ApiOut<MusicDetailOut>> => {
   return await apiOut(apiClient.get(apiMusic(id), cookieHeader(req)))
 }
 
-export const getComic = async (id: number, req?: Req): Promise<ApiOut<Comic>> => {
+export const getComic = async (id: number, req?: Req): Promise<ApiOut<ComicDetailOut>> => {
   return await apiOut(apiClient.get(apiComic(id), cookieHeader(req)))
 }
 
-export const getPicture = async (id: number, req?: Req): Promise<ApiOut<Picture>> => {
+export const getPicture = async (id: number, req?: Req): Promise<ApiOut<PictureDetailOut>> => {
   return await apiOut(apiClient.get(apiPicture(id), cookieHeader(req)))
 }
 
@@ -26,10 +25,6 @@ export const getBlog = async (id: number, req?: Req): Promise<ApiOut<BlogDetailO
   return await apiOut(apiClient.get(apiBlog(id), cookieHeader(req)))
 }
 
-export const getChat = async (id: number, req?: Req): Promise<ApiOut<Chat>> => {
+export const getChat = async (id: number, req?: Req): Promise<ApiOut<ChatDetailOut>> => {
   return await apiOut(apiClient.get(apiChat(id), cookieHeader(req)))
-}
-
-export const postFollow = async (request: FollowIn): Promise<ApiOut<FollowOut>> => {
-  return await apiOut(apiClient.post(apiFollow, camelSnake(request)))
 }
