+++ frontend/src/components/widgets/Media/Detail/Common.tsx
@@ -1,10 +1,11 @@
 import { ChangeEvent, useState, useEffect, useMemo } from 'react'
 import { useRouter } from 'next/router'
 import clsx from 'clsx'
+import { FollowIn } from 'types/internal/auth'
 import { Comment, CommnetIn } from 'types/internal/comment'
-import { Author, FollowIn, MediaUser } from 'types/internal/media'
+import { Author, MediaUser } from 'types/internal/media'
 import { postComment } from 'api/internal/media/comment'
-import { postFollow } from 'api/internal/media/detail'
+import { postFollow } from 'api/internal/user'
 import { CommentType, FetchError } from 'utils/constants/enum'
 import { commentTypeNoMap } from 'utils/constants/map'
 import { capitalize, isActive } from 'utils/functions/common'
