+++ frontend/src/types/internal/media/detail.d.ts
@@ -1,0 +1,80 @@
+import { Comment } from 'types/internal/comment'
+import { Author, MediaUser } from 'types/internal/media'
+import { Video, Music, Comic, Picture, Blog, Chat } from 'types/internal/media'
+
+export interface MediaDetail {
+  id: number
+  title: string
+  content: string
+  like: number
+  read: number
+  publish: boolean
+  created: Date
+  updated: Date
+  author: Author
+  mediaUser: MediaUser
+}
+
+export interface VideoDetail extends Video {
+  comments: Comment[]
+  hashtags: string[]
+}
+
+export interface MusicDetail extends Music {
+  comments: Comment[]
+  hashtags: string[]
+}
+
+export interface ComicDetail extends Comic {
+  comments: Comment[]
+  hashtags: string[]
+}
+
+export interface PictureDetail extends Picture {
+  comments: Comment[]
+  hashtags: string[]
+}
+
+export interface BlogDetail extends MediaDetail {
+  image: str
+  richtext: str
+  comments: Comment[]
+  hashtags: string[]
+}
+
+export interface ChatDetail extends MediaDetail {
+  read: number
+  joined: number
+  thread: number
+  hashtags: string[]
+}
+
+export interface VideoDetailOut {
+  detail: VideoDetail
+  list: Video[]
+}
+
+export interface MusicDetailOut {
+  detail: MusicDetail
+  list: Music[]
+}
+
+export interface ComicDetailOut {
+  detail: ComicDetail
+  list: Comic[]
+}
+
+export interface PictureDetailOut {
+  detail: PictureDetail
+  list: Picture[]
+}
+
+export interface BlogDetailOut {
+  detail: BlogDetail
+  list: Blog[]
+}
+
+export interface ChatDetailOut {
+  detail: ChatDetail
+  list: Chat[]
+}
